+++ subsys/bluetooth/host/audio/CMakeLists.txt
@@ -15,6 +15,7 @@ zephyr_library_sources_ifdef(CONFIG_BT_TBS tbs.c)
 zephyr_library_sources_ifdef(CONFIG_BT_CCP ccp.c)
 zephyr_library_sources_ifdef(CONFIG_BT_PACS pacs.c)
 zephyr_library_sources_ifdef(CONFIG_BT_AUDIO_CHAN chan.c)
+zephyr_library_sources_ifdef(CONFIG_BT_AUDIO_BROADCAST broadcast_source.c)
 zephyr_library_sources_ifdef(CONFIG_BT_AUDIO_ENDPOINT endpoint.c)
 zephyr_library_sources_ifdef(CONFIG_BT_AUDIO_CAPABILITIES capabilities.c)
 zephyr_library_sources_ifdef(CONFIG_BT_ASCS ascs.c)
